Endian fix for reading DS2408 set_alarm property
Writing set_alarm sets the proper alarm on the device, but trying to
read it back yielded invalid response. Example:

Writing 10000000, readback gave 20000000
Writing 133333333, readback gave 122222230
Writing 311111111, readback gave 322222222

Note: has this worked before? If so, on what arch was it tested?
This fixed it on x86 platform (little-endian), but might break on
others?