Mutex fixes from Stefano Miccoli
my owserver (current development git version) crashes on OSX when one
tries to read /settings/alias/list. (This does not happens on Linux).

I found out that this is due to an uninitialized member in the global
struct Mutex, see module/owlib/src/include/ow_mutexes.h On the hunt for
similar errors I found other struct Mutex members that are not
initialized and/or not used.

The attached patch fixes this problem.

Actions are

1) init Mutex.aliaslist_mutex
2) delete reference to unused Mutex.simul_mutex, Mutex.aliasfind_mutex
3) correct a fprintf call from  fprintf(stderr, buf) to fprintf(stderr,
"%s", buf). This latter correction may seem unnecessary, but the CLANG
compiler flags fprintf(stderr, buf) as a security problem, and I think
that although the injection of malicious code into 'buf' is very
unlikely, it is always better to stay on the safe side.

Regards,

Stefano