Added support to interact with LINK 12/45/USB AUX line
This is a special physical line on the LINK adapter which can be used
as a general purpouse I/O pin. It can either be in output mode, or in
input mode (high-z), but not both at the same time (i.e. no way to read
current setting).

Putting it in ouput mode is done by writing 1 (high) or 0 (low)
to /bus.0/interface/settings/link/auxout. Readback is not supported, it
would be a guess by emitting the last written value.

Putting in high-Z mode and reading current line level is done by
reading /bus.0/interface/settings/link/auxin.