compat.h: only include compat_netdb.h if needed
compat_netdb.h is added for only one reason: to provide the definitions
for getaddrinfo, in case it is not provided by the system.
compat_netdb.h is just a copy from an old glibc version, but that makes
it incompatible with other standard C libraries. For example, uClibc
without RPC support doesn't have the rpc/netdb.h header. This leads to
build errors.

To work around this, only include compat_netdb.h when needed, i.e.
when getaddrinfo isn't provided natively.

Note that this getaddrinfo compat was introduced in commit bc54c431 to
support Solaris 7. Starting from Solaris 8, however, getaddrinfo is
available natively, and it's available on MacOS, Windows and any other
POSIX-2001 compliant system as well. So it may be worhtwhile to remove
it altogether.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Johan Ström <johan@stromnet.se>