changed some 'LINK' to ''LinkUSB'
Initially I was confused, if 'LINK' was the same as the 'LinkUSB', especially because sometimes indeed it was called 'LinkUSB'.
So I think naming it by it's more specific name 'LinkUSB' is more clear.
Just a minor change, to see how things are going. I am not an experienced github user.