Fix race conditions in ow_regcomp
The following trivial program will occasionally SIGSEGV, SIGABRT, or
deadlock:

  #include <owcapi.h>
  int main() {
    char* buf;
    size_t len;
    OW_init("--fake 20 --fake 42");
    OW_get("/",&buf,&len);
    return 0;
  }

Backtraces show that these occur within calls to ow_regcomp, and most
often during the underlying regcomp(3) call. It is also possible during
the tsearch(3) call, which is explicitly thread-unsafe against a shared
root.

This patch replaces the tsearch check with the standard pthread_once
mechanism, which is designed specifically for this use case.

Signed-off-by: Justin Brewer <justin.brewer@vertivco.com>
Signed-off-by: Johan Ström <johan@stromnet.se>