Add pkg-config entry for owcapi
This allows downstream packages to use PKG_CHECK_MODULES to pull in
owcapi compilation and linking flags, rather than hardcoding them.

According to the pkg-config version parsing rules, the version "3.2p2"
will not be parsed as "3 2 2", but as "3 2 p 2" (the p is considered the
third component of the version). Fortunately, when comparing numeric and
alphabetic components, pkg-config will consider the numeric version
newer. So, if in the future, the 'p' is replaced with '.', version
ordering will remain consistent.

A backport of PKG_INSTALLDIR is provided, since it's a relatively new
addition. By relatively new, I mean it was added 6 years ago, but
there's still plenty of build hosts running older versions.

Signed-off-by: Justin Brewer <justin.brewer@vertivco.com>