Add support for filtering devices with family code
Each 1-wire slave-device has their own 64bits serial number. The
first 8 bits of that serial number is used to define the family of the
device (temperature sensor, button, eprom, ...).
This commit adds a new path type for filtering the devices on a
directory with their family code.

For instance the following url for owhttpd or query in owlib are now
supported:

  - /bus.0/bus.2/family.01/
  - /uncached/bus.0/family.2A/

The next commit will implement for different 1-wire chip backend
an improved search algorithm that speeds up the discovery of 1-wire
devices for a single family code.

For unsupported backends the filtering is done once the device get
discovered. Backends not supporting the family search still have to
discover every devices on the bus. Thus, no speedup is gained for them,
but the behavior stays the same.