Define 1-wire device serial number bit length
The SN bit-length is used in the code base by the discovery algorithm
and test device presence.
On certain part of the code base `SERIAL_NUMBER_BITS` is defined
and other a magic number is used.
This commit replace the usage of magic numbers in the code and move
`SERIAL_NUMBER_BITS` to the same place as `SERIAL_NUMBER_SIZE`.